+++ assets/js/atomic/blocks/product-elements/shared/config.js
@@ -7,7 +7,7 @@ import { isExperimentalBuild } from '@woocommerce/block-settings';
 /**
  * Internal dependencies
  */
-import save from '../../save';
+import save from '../save';
 
 /**
  * Holds default config for this collection of blocks.
